+++ shared/redis-client.js
